yeti-vibes/source/requirements.txt
absl-py==2.1.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
attrs==24.2.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.5
decorator==5.1.1
Django==5.1.1
django-environ==0.11.2
djangorestframework==3.15.2
exceptiongroup==1.2.2
executing==2.0.1
filelock==3.13.1
flatbuffers==24.3.25
fonttools==4.53.1
fsspec==2024.2.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.65.4
h5py==3.11.0
idna==3.7
ipykernel==6.29.5
ipython==8.26.0
jax==0.4.31
jaxlib==0.4.31
jedi==0.19.1
Jinja2==3.1.3
jupyter_client==8.6.2
jupyter_core==5.7.2
keras==3.5.0
kiwisolver==1.4.5
lapx==0.5.9.post1
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1.post1
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.3
ml-dtypes==0.4.0
mpmath==1.3.0
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.3
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pandas==2.2.2
parso==0.8.4
pillow==10.2.0
pip-review==1.3.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
protobuf==3.20.3
psutil==6.0.0
psycopg2-binary==2.9.9
pure_eval==0.2.3
py-cpuinfo==9.0.0
pycparser==2.22
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32==306
PyYAML==6.0.2
pyzmq==26.1.0
requests==2.32.3
rich==13.7.1
scipy==1.14.0
seaborn==0.13.2
shapely==2.0.5
six==1.16.0
sounddevice==0.5.0
sqlparse==0.5.1
stack-data==0.6.3
sympy==1.12
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
torch==2.4.0+cu121
torchaudio==2.4.0+cu121
torchvision==0.19.0+cu121
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.9.0
tzdata==2024.1
ultralytics==8.2.74
ultralytics-thop==2.0.0
urllib3==2.2.2
wcwidth==0.2.13
Werkzeug==3.0.3
wrapt==1.16.0
